Add Data aggregation Engine
Add a singleton class `Aggregator` to handle all data aggregation
structure and with the capabilities to send data to astarte.
Add a `Report` Class to model the report message.
Add two `ScheduledTask` classes to handle the periodic send of
aggregated data.
`ScheduledGroupTask` is currently TBI, waiting for the Astarte Java SDK
update.

Signed-off-by: Francesco Vaiani <francesco.vaiani@secomind.com>